requirements.txt
streamlit
tensorflow-cpu
pandas
Pillow
numpy
plotly
streamlit-option-menu
fpdf2
opencv-python-headless
